refactoring resourcesDialog to be in a separate component
